pyproject.toml
[tool.poetry]
name = "fastapi-auth"
version = "0.1.0"
description = ""
authors = ["harmelikyan92 <harut@playbrush.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bcrypt = "^4.1.2"
fastapi = "^0.109.2"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.6.1"}
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.27"
uvicorn = "^0.27.1"
pip = "^24.0"
install = "^1.3.5"
psycopg2 = "^2.9.9"
prometheus-fastapi-instrumentator = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
